changes in homepage, header and footer

+++ frontend/src/components/service/UserService.js
@@ -3,10 +3,12 @@ import axios from "axios";
 class UserService {
     static BASE_URL = "http://localhost:8080";
 
+
     static async login(email, password){
         try {
-
-            const response = await axios.post(`${UserService.BASE_URL}/auth/login`, {email, password})
+            const response = await axios.post(`${UserService.BASE_URL}/auth/login`, { email, password })
+            const {token, role} = response.data;
+            UserService.saveData(token, role);  
             return response.data;
 
         } catch (err) {
@@ -16,7 +18,6 @@ class UserService {
 
     static async registerAdmin(userData){
         try {
-
             const response = await axios.post(`${UserService.BASE_URL}/auth/register`, userData,)
             return response.data;
 
@@ -39,7 +40,7 @@ class UserService {
 
     static async sendOtp(email){
         try {
-            const response = await axios.post(`${UserService.BASE_URL}/auth/send-otp`, {email})
+            const response = await axios.post(`${UserService.BASE_URL}/auth/send-otp`, { email })
             return response.data;
             
         } catch (err) {
@@ -50,7 +51,7 @@ class UserService {
 
     static async verifyOtp(email, otp){
         try{
-            const response = await axios.post(`${UserService.BASE_URL}/verify-otp`, {email, otp})
+            const response = await axios.post(`${UserService.BASE_URL}/verify-otp`, { email, otp })
             return response.data;
         } 
         catch (err){
@@ -60,7 +61,7 @@ class UserService {
 
     static async updatePassword(email, password){
         try{
-            const response = await axios.put(`${UserService.BASE_URL}/auth/update-password`, {email, password})
+            const response = await axios.put(`${UserService.BASE_URL}/auth/update-password`, { email, password })
             return response.data;
         } 
         catch (err){
@@ -70,11 +71,12 @@ class UserService {
 
     static async getAllUsers(token){
         try {
-
+            const token = UserService.getToken();
+            if (!token) {throw new Error("No se encontró token")}
             const response = await axios.get(`${UserService.BASE_URL}/admin/get-all-users`,
                 {
                     headers: {Authorization: `Bearer ${token}`}
-        })
+        });
             return response.data;
 
         } catch (err) {
@@ -84,11 +86,12 @@ class UserService {
 
     static async getYourProfile(token){
         try {
-
+            const token = UserService.getToken();
+            if (!token) {throw new Error("No se encontró token")}
             const response = await axios.get(`${UserService.BASE_URL}/adminuser/get-profile`,
                 {
                     headers: {Authorization: `Bearer ${token}`}
-        })
+        });
             return response.data;
 
         } catch (err) {
@@ -98,11 +101,12 @@ class UserService {
 
     static async getUserById(userId, token){
         try {
-
+            const token = UserService.getToken();
+            if (!token) {throw new Error("No se encontró token")}
             const response = await axios.get(`${UserService.BASE_URL}/admin/get-user/${userId}`,
                 {
                     headers: {Authorization: `Bearer ${token}`}
-        })
+        });
             return response.data;
 
         } catch (err) {
@@ -112,11 +116,12 @@ class UserService {
 
     static async deleteUser(userId, token){
         try {
-
+            const token = UserService.getToken();
+            if (!token) {throw new Error("No se encontró token")}
             const response = await axios.delete(`${UserService.BASE_URL}/admin/delete/${userId}`,
                 {
                     headers: {Authorization: `Bearer ${token}`}
-        })
+        });
             return response.data;
 
         } catch (err) {
@@ -126,11 +131,12 @@ class UserService {
 
     static async updateUser(userId, userData, token){
         try {
-
+            const token = UserService.getToken();
+            if (!token) {throw new Error("No se encontró token")}
             const response = await axios.put(`${UserService.BASE_URL}/admin/update/${userId}`, userData,
                 {
                     headers: {Authorization: `Bearer ${token}`}
-        })
+        });
             return response.data;
 
         } catch (err) {
@@ -140,28 +146,34 @@ class UserService {
 
     //** AUTHENTICATION CHECKER */
 
+    static async saveData(token, role) {
+        localStorage.setItem('token', token);
+        localStorage.setItem('role', role);
+    }
+
     static logout(){
         localStorage.removeItem('token')
         localStorage.removeItem('role')
     }
 
+    static getToken(){
+        return localStorage.getItem('token');
+    }
+
+    static getRole(){
+        return localStorage.getItem('role');
+    }
+
     static isAuthenticated(){
-        const token = localStorage.getItem('token')
-        return !!token
+        return !!UserService.getToken();
     }
 
     static isAdmin(){
-        const role = localStorage.getItem('role')
-        return role === 'ADMIN'
+        return UserService.isAuthenticated && UserService.getRole() === 'ADMIN';   
     }
 
     static isUser(){
-        const role = localStorage.getItem('role')
-        return role === 'USER'
-    }
-
-    static adminOnly(){
-        return this.isAuthenticated() && this.isAdmin();
+        return UserService.isAuthenticated && UserService.getRole() === 'USER';
     }
 }
